fix: adding update quantity in pending orders. UI changes. Changing quantity from double to int

+++ lib/core/data/local_storage/models/hive_cart_model.dart
@@ -17,7 +17,7 @@ class HiveCartModel extends HiveObject {
   @HiveField(4)
   String productName;
   @HiveField(5)
-  double quantity;
+  int quantity;
   @HiveField(6)
   double rate;
   @HiveField(7)
